workshop/requirements.txt
blosc
diskcache
h5py
jupyter
matplotlib
numpy
Pillow
scikit-image
scipy
SimpleITK
torch